+++ backend/purchasing/ports/po_repo_port.py
@@ -28,7 +28,7 @@ class PORepoPort(ABC):
         item_id: str,
         status: POItemStatus,
         product_id: Optional[str] = None,
-        delivered_qty: Optional[int] = None,
+        delivered_qty: Optional[float] = None,
     ) -> None: ...
 
     @abstractmethod
